+++ src/Model/PaymentMethodRestrictionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace MangoSylius\PaymentRestrictionPlugin\Model;
+namespace ThreeBRS\SyliusPaymentRestrictionPlugin\Model;
 
 use Doctrine\Common\Collections\Collection;
 use Sylius\Component\Addressing\Model\ZoneInterface;