+++ src/Model/PaymentMethodRestrictionTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace MangoSylius\PaymentRestrictionPlugin\Model;
+namespace ThreeBRS\SyliusPaymentRestrictionPlugin\Model;
 
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
@@ -20,7 +20,7 @@ trait PaymentMethodRestrictionTrait
 	/**
 	 * @var Collection<array-key, ShippingMethodInterface>|ShippingMethodInterface[]
 	 * @ORM\ManyToMany(targetEntity="Sylius\Component\Core\Model\ShippingMethod", inversedBy="paymentMethods")
-	 * @ORM\JoinTable(name="mangoweb_payment_method_shipping_method",
+	 * @ORM\JoinTable(name="threebrs_payment_method_shipping_method",
 	 *     joinColumns={@ORM\JoinColumn(name="payment_method_id", referencedColumnName="id")},
 	 *     inverseJoinColumns={@ORM\JoinColumn(name="method_shipping_id", referencedColumnName="id")}
 	 * )