+++ src/ThreeBRSSyliusPaymentRestrictionPlugin.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace MangoSylius\PaymentRestrictionPlugin;
+namespace ThreeBRS\SyliusPaymentRestrictionPlugin;
 
 use Sylius\Bundle\CoreBundle\Application\SyliusPluginTrait;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class MangoSyliusPaymentRestrictionPlugin extends Bundle
+class ThreeBRSSyliusPaymentRestrictionPlugin extends Bundle
 {
 	use SyliusPluginTrait;
 }