+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v2.0.0 (2021-11-09)
+
+#### Details
+
+- Support for Sylius 1.8|1.9|1.10, Symfony ^4.4|^5.2, PHP ^7.3|^8.0
+- Change namespace from `MangoSylius\PaymentRestrictionPlugin` to `ThreeBRS\SyliusPaymentRestrictionPlugin`
+- Change table name from `mango_payment_method_shipping_method` to `threebrs_payment_method_shipping_method`
+- *BC break: Version 3.0.0 is NOT compatible with previous versions due to namespace change*
+
 ## v1.0.0 (2020-06-17)
 
 #### Details